requirements.txt
Flask~=2.3.2
google-api-python-client
requests~=2.31.0
openai~=1.8.0
python-dotenv~=1.0.0
flask_session
flask_cors
numpy
langdetect~=1.0.9
SQLAlchemy==2.0.12
bcrypt~=4.0.1
flask_sqlalchemy
flask_migrate
psycopg2-binary
translate
ipython
alembic~=1.12.0
googletrans==3.1.0a0
stripe~=7.13.0
werkzeug~=3.0.1